+++ Modules/Intent.Modules.ApplicationTemplate.Builder/release-notes.md
@@ -1,4 +1,9 @@
-﻿### Version 3.6.1
+﻿### Version 3.7.0
+
+- Improvement: Added support for `Is New`, `Incompatibilities`, `Documentation Url` and `Tags` to components, which will be respected in version `4.5.23` of Intent Architect and later.
+- Improvement: Icons for components will now render in the designer.
+
+### Version 3.6.1
 
 - Feature: A context menu option has been added to packages to update versions from the content of a modules.config file.