+++ Modules/Intent.Modules.Metadata.Security/release-notes.md
@@ -1,3 +1,8 @@
+### Version 1.1.2
+
+- Improvement: Updated module documentation to use centralized documentation site.
+- Fixed: Removed `Secure X` suggestions from End points which come from other applications.
+
 ### Version 1.1.1
 
 - Fixed: Removed duplicate validation being done on Commands, Queries and Operations.